+++ server/src/net/stzups/tanks/game/Player.java
@@ -3,7 +3,7 @@ package net.stzups.tanks.game;
 class Player extends Object {
     private String name;
 
-    Player(String name, int x, int y) {
+    Player(String name, double x, double y) {
         super(x, y, 0, 0);
         this.name = name;
     }
